complete rewrite
Completely rewrite how messages are handled and some other (minor)
changes as well.

This rewrite allows to have a single file per supported property. So
just having a file / bridge to for example have the power handling. So
when adding a new feature in can just be contained in a single file
instead of having to modify the CEC handler, creating messages, etc.
This greatly simplifies things while also making it more flexible and
guture proof.