config: sync new device options to config
When there was no exact match for a device but there was only one vlid
option sync back the address, vendor, name etc of the device back to
this only option. This makes sure that the correct "settings" are saved
again for the device so the next time there hopefully will be a full
match again.