power: use message argument when updating power status
When updating the power status based on the Report Power Status message
directly use the value passed in the message. This as there is a big
chance that using libcecs GetPowerStatus will make (another) call to get
the power status as it hasn't handled the reply (which we're using by
using the log callback) yet.