travis: optimize build by utilizing the build matrix
Using the build matrix it is possible to compile for the different
architectures and platforms in parallel, so utilize that to speed up the
build. This as building all at once in a single job resulted in a build
time of more thatn 45 minutes which thus also would (potentially) be
killed by the timeout of 50 seconds.

Besides building the containers in parallel the build process has also
been optimized as there is now a tag (per architecture + platform) for
after the cec build phase which is reused in later builds. This saves
about 10 minutes of build time in the arm builds as the CEC library
doesn't have to be recompiled anmore (as long as the base image and
libcec didn't chnage).